+++ src/QuantitativeWorld.Tests/EnumerableExtensionsTests.SumLinearQuantitiesWithFactoryFunctionBySelector.cs
@@ -19,7 +19,7 @@ namespace QuantitativeWorld.Tests
             {
                 // arrange
                 IEnumerable<TestObject<Power>> objects = null;
-                Func<decimal, PowerUnit, Power> factory = PowerFactory.Create;
+                Func<double, PowerUnit, Power> factory = PowerFactory.Create;
                 Func<TestObject<Power>, Power> selector = e => e.Property;
 
                 // act
@@ -35,7 +35,7 @@ namespace QuantitativeWorld.Tests
             {
                 // arrange
                 var objects = Enumerable.Empty<TestObject<Power>>();
-                Func<decimal, PowerUnit, Power> factory = PowerFactory.Create;
+                Func<double, PowerUnit, Power> factory = PowerFactory.Create;
                 Func<TestObject<Power>, Power> selector = null;
 
                 // act
@@ -51,7 +51,7 @@ namespace QuantitativeWorld.Tests
             {
                 // arrange
                 var objects = Enumerable.Empty<TestObject<Power>>();
-                Func<decimal, PowerUnit, Power> factory = null;
+                Func<double, PowerUnit, Power> factory = null;
                 Func<TestObject<Power>, Power> selector = e => e.Property;
 
                 // act
@@ -67,7 +67,7 @@ namespace QuantitativeWorld.Tests
             {
                 // arrange
                 var objects = Enumerable.Empty<TestObject<Power>>();
-                Func<decimal, PowerUnit, Power> factory = PowerFactory.Create;
+                Func<double, PowerUnit, Power> factory = PowerFactory.Create;
                 Func<TestObject<Power>, Power> selector = e => e.Property;
 
                 // act
@@ -82,10 +82,10 @@ namespace QuantitativeWorld.Tests
             {
                 // arrange
                 var objects = Fixture.CreateMany<Power>(3).Select(e => new TestObject<Power>(e));
-                Func<decimal, PowerUnit, Power> factory = PowerFactory.Create;
+                Func<double, PowerUnit, Power> factory = PowerFactory.Create;
                 Func<TestObject<Power>, Power> selector = e => e.Property;
 
-                decimal expectedResultInWatts = objects.Sum(e => e.Property.Value * e.Property.Unit.ValueInWatts);
+                double expectedResultInWatts = objects.Sum(e => e.Property.Value * e.Property.Unit.ValueInWatts);
                 var expectedResultUnit = objects.First().Property.Unit;
                 var expectedResult = new Power(expectedResultInWatts / expectedResultUnit.ValueInWatts, expectedResultUnit);
