+++ src/QuantitativeWorld/PowerFactory.cs
@@ -4,7 +4,7 @@ namespace QuantitativeWorld
 {
     public class PowerFactory : ILinearQuantityFactory<Power, PowerUnit>
     {
-        public Power Create(decimal value, PowerUnit unit) =>
+        public Power Create(double value, PowerUnit unit) =>
             new Power(value, unit);
     }
 }
