+++ src/QuantitativeWorld/WeightUnit.cs
@@ -7,13 +7,13 @@ namespace QuantitativeWorld
     {
         private readonly string _name;
         private readonly string _abbreviation;
-        private readonly decimal? _valueInKilograms;
+        private readonly double? _valueInKilograms;
 
-        public WeightUnit(string name, string abbreviation, decimal valueInKilograms)
+        public WeightUnit(string name, string abbreviation, double valueInKilograms)
         {
             Assert.IsNotNullOrWhiteSpace(name, nameof(name));
             Assert.IsNotNullOrWhiteSpace(abbreviation, nameof(abbreviation));
-            Assert.IsGreaterThan(valueInKilograms, 0m, nameof(valueInKilograms));
+            Assert.IsGreaterThan(valueInKilograms, 0d, nameof(valueInKilograms));
 
             _name = name;
             _abbreviation = abbreviation;
@@ -22,10 +22,10 @@ namespace QuantitativeWorld
 
         public string Name => _name ?? Kilogram._name;
         public string Abbreviation => _abbreviation ?? Kilogram._abbreviation;
-        public decimal ValueInKilograms => _valueInKilograms ?? Kilogram._valueInKilograms.Value;
+        public double ValueInKilograms => _valueInKilograms ?? Kilogram._valueInKilograms.Value;
 
         public override string ToString() => Abbreviation;
 
-        decimal ILinearUnit.ValueInBaseUnit => ValueInKilograms;
+        double ILinearUnit.ValueInBaseUnit => ValueInKilograms;
     }
 }
